requirements.txt
PyQt6>=6.5.0
opencv-python>=4.8.0
python-docx>=1.1.0
pyinstaller>=6.0.0